% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapus_adhb_konstruksi.R
\name{konstruksi_adhb}
\alias{konstruksi_adhb}
\title{Memuat lapangan usaha F. Konstruksi}
\usage{
konstruksi_adhb()
}
\description{
Fungsi ini menjalankan alur penghitungan PDRB menurut lapangan usaha konstruksi.
}
\examples{
konstruksi_adhb()
}
