% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapus_adhk_konstruksi.R
\name{konstruksi_adhk}
\alias{konstruksi_adhk}
\title{Memuat lapangan usaha F. Konstruksi}
\usage{
konstruksi_adhk()
}
\description{
Fungsi ini menjalankan alur penghitungan PDRB menurut lapangan usaha konstruksi.
}
\examples{
konstruksi_adhk()
}
